Rename packages in :groovy-loader to match the project structure
The `classloading` name was too generic.
