+++ platforms/core-runtime/groovy-loader/src/main/java/org/gradle/internal/groovyloader/PreferenceCleaningGroovySystemLoader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.classloading;
+package org.gradle.internal.groovyloader;
 
 import org.gradle.api.GradleException;
