+++ subprojects/core/src/main/java/org/gradle/api/internal/project/antbuilder/ClassPathToClassLoaderCache.java
@@ -16,12 +16,12 @@
 package org.gradle.api.internal.project.antbuilder;
 
 import org.gradle.api.Action;
-import org.gradle.api.internal.classloading.GroovySystemLoader;
-import org.gradle.api.internal.classloading.GroovySystemLoaderFactory;
 import org.gradle.internal.Factory;
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.classpath.ClassPath;
 import org.gradle.internal.concurrent.Stoppable;
+import org.gradle.internal.groovyloader.GroovySystemLoader;
+import org.gradle.internal.groovyloader.GroovySystemLoaderFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
