+++ subprojects/core/src/main/java/org/gradle/api/internal/project/antbuilder/Cleanup.java
@@ -15,9 +15,9 @@
  */
 package org.gradle.api.internal.project.antbuilder;
 
-import org.gradle.api.internal.classloading.GroovySystemLoader;
 import org.gradle.internal.classloader.ClassLoaderUtils;
 import org.gradle.internal.classpath.ClassPath;
+import org.gradle.internal.groovyloader.GroovySystemLoader;
 
 import java.lang.ref.PhantomReference;
 import java.lang.ref.ReferenceQueue;
